Fixing bug 659347: User complained that the excuse module is missing.
The excuse server itself has stopped working, so I settled for
simply pulling down the complete list of excuses and using that,
rather like the zippy module does.  Since the excuse server is a
circa-1996 alt.sysadmin.recovery in-joke, and has long since been
forgotten by that crowd, I figure it's not likely to be updated
that often.