Dammit.
I found this one in the first place.  It was posted to
uw.mfcf.questions or some such, ages and ages ago, and I passed it
on to alt.sysadmin.recovery, and everyone had a good laugh at it
and it made it into various standard BOFH excuse lists.

"The computer fletely, mouse and all."

Sheesh.  At least get it right.